+++ NEWS.md
@@ -5,6 +5,10 @@
 * `wav_2_flac()` to convert '.wav' files into flac compresed lossless format
 * `gaps()` for calculating time intervals (i.e. gaps) between selections
 * `read_sound_files()` is now use to read sound files
+
+## Changes and additions:
+
+- `autodetec()`: if 'maxdur' is supplied then merging selections through 'hold.time' will only applied when the new merged selection's duration is equal or lower than 'maxdur'.
 * warbleR now supports '.wav', '.flac', '.wac' and '.mp3' formats
 
 # *warbleR 1.1.26*